separete rdb
