+++ backend/internal/config/rdb.go
@@ -1,0 +1,16 @@
+package config
+
+type RelationalDB struct {
+	Master   DB   `yaml:"master"`
+	Repricas []DB `yaml:"repricas"`
+}
+
+type DB struct {
+	Driver   string `yaml:"driver"`
+	Host     string `yaml:"host"`
+	Port     string `yaml:"port"`
+	User     string `yaml:"user"`
+	Name     string `yaml:"name"`
+	Password string `yaml:"password"`
+	Sslmode  string `yaml:"sslMode"`
+}
